cv/src/rtdetrv2_pytorch/requirements.txt
torch>=2.0.1
torchvision>=0.17.0
faster-coco-eval>=1.6.5
PyYAML
tensorboard
scipy
pycocotools
onnx
onnxruntime-gpu
tensorrt==8.5.2.2